+++ R/crs.R
@@ -46,9 +46,14 @@ scalefactor_lcc <- function(par_deg, par1_deg, par2_deg) {
 #' Calculate the cone constant for a normal LCC projection
 #'
 #' @keywords internal
+#' @noRd
 coneconst_lcc <- function(par1, par2) {
-  log(cos(par1) / cos(par2)) / log(tan(pi / 4 + par2 / 2) /
-    tan(pi / 4 + par1 / 2))
+  log(
+    cos(par1) / cos(par2)
+  ) /
+    log(
+      tan(pi / 4 + par2 / 2) / tan(pi / 4 + par1 / 2)
+    )
 }